flask_app/requirements.txt
alabaster==0.7.12
alembic==1.5.2
argh==0.26.2
astropy==4.0.2
async-generator==1.10
atomicwrites==1.4.0
awscli==1.19.5
babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bkcharts==0.2
boto==2.49.0
botocore==1.20.5
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
ecdsa==0.14.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
glob2==0.7
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
imagesize==1.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jupyter==1.0.0
jupyter-core==4.6.3
jupyterlab==2.2.6
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
Mako==1.1.4
mccabe==0.6.1
mock==4.0.2
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
numexpr==2.7.1
olefile==0.46
packaging==20.4
parso==0.7.0
partd==1.1.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycosat==0.6.3
pycurl==7.43.0.6
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtPy==1.9.0
rsa==4.5
ruamel-yaml==0.15.87
s3transfer==0.3.4
scikit-image==0.17.2
Send2Trash==1.5.0
simplegeneric==0.8.1
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.22
tables==3.6.1
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
unicodecsv==0.14.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.3.3
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zope.event==4.5.0